[Refactor]: 템플릿에서 currentPage 저장하도록
